backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "bcrypt>=4.3.0",
    "email-validator>=2.0.0",
    "fastapi>=0.115.13",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "psycopg[async,binary]>=3.2.9",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
